+++ printcetnrum_web.client/src/app/file-pages/diploma-page/diploma-page.component.css
