+++ printcetnrum_web.client/src/app/file-pages/stamp-page/stamp-page.component.spec.ts
