+++ printcetnrum_web.client/src/app/home-page/home-page.component.css
@@ -9,7 +9,7 @@ header.masthead {
   padding-bottom: 6rem;
   text-align: center;
   color: #fff;
-  background-image: url("../../../assets/about-pictures/PrintCentrum_inside.jpg");
+  background-image: url("../../assets/about-pictures/PrintCentrum_inside.jpg");
   background-repeat: no-repeat;
   background-attachment: scroll;
   background-position: center center;
